Fix mask error

+++ moviepy/video/VideoClip.py
@@ -611,7 +611,7 @@ class VideoClip(Clip):
         im_img = Image.fromarray(img)
 
         if self.mask is not None:
-            mask = self.mask.get_frame(ct)
+            mask = self.mask.get_frame(ct).astype("uint8")
             im_mask = Image.fromarray(255 * mask).convert("L")
 
             if im_img.size != im_mask.size:
